Add files via upload
added initial support for the message passing interface in the strassen and winograd algorithms. this is, however, entirely untested. Will move to the main folder if tested. 
Unfortunately, I still have to use new and delete in the gpu functions as nvc++'s linker can not find acc_alloc and acc_free.