Update mdspan_acc.h
the openacc pragmas are now such that nvc++ can vectorize more loops (mostly an independent clause had to be added). Unfortunately, openmp has no "independent" clause for loops. nvc++ is unfortunately confused by the matrix strides for the openmp host code and refuses to vectorize them. this seems to be different from other compilers, like gcc or clang, which can vectorize such code. Unfortunately, gcc and clang have difficulties with the open-acc and openmp offload by now, which is still in development for gcc and clang.