Update mdspan_omp.h
now the Lu/cholesky and QR decomposition use mostly teams of threads. Support for unified shared memory was added but not tested. Full support for rather complete parallelization is only possible with shared memory, due to openmp's restriction on the teams distribute keyword. One can now use the library with -O3 on clang.