Update mdspan_omp.h
Now the cholesky/Lu and QR decomposition mostly use teams of threads. Initial support for shared memory was added but not tested. Especially for the qr decomposition, full gpu parallelization seems only possible with shared memory, due to openmp's restrictions on the teams keyword. The library now compiles with -O3 with clang.