Update mdspan_omp.h
fixes in the shared clauses of some openmp statements.
The send, recieve and broadcast functions of the message passing interface now work with the library with the mdspan class and the datastruct structure. I have put the recieve into the constructor and made send a memberfunction of the mdspan class. Unfortunately, the strassen and winograd algorithms still have some problems with the message passing interface and currently only work on OpenMP