Add files via upload
fixed two errors in the gpu memory functions. One variable was stored as copy and not as reference. This confused the data offload. in the update function, the data field was updated instead of the strides... Fixed the strassen and winograd algorithms, which can now work entirely on device with datapointers on device