Add files via upload
In this version, the advanced algorithms now work on gpu. The QR decomposition has some changes made when compared to https://arxiv.org/pdf/1812.02056 for increased numerical stability. Unfortunately, because of apparent miscompilation, the advanced algorithm for the Cholesky decomposition does not seem to work on gpu properly because of https://gcc.gnu.org/bugzilla/show_bug.cgi?id=121818