Add files via upload
Separated the various demonstrations and tests of the classes into several programs. Added a nice syntax for expression manipulation. Then, one can write expressions like C=A*B, Added some speed optimizations for the library with const arguments, added constructors for vectors, changed the order of arguments of the constructors in mdspan to make it more usable for humans. The expressions C=A*B now use policy objects to enquire whether they should work on gpu, This makes it sower than direct calls. The mechanism does not provide the chaining of operations yet. Impmenenting the ability to write C=A*(B+C..) would be a mere line. The problem is it would require the storage of intermediate objects. While the library provides a class which can do this, the problem is that if A is 1 GB large, the library would then automatically have to create large intermediate matrices on gpu or cpu or harddrive, and the user would give up control . I am still thinking whether to implement something like this.