Add files via upload
Optimizations for the Strassen and Winograd Algorithm. Added more support for operations with sparse matrices but i have not tested this by now.

After clang fixed some bugs in version 21.1.2, the library now compiles and runs with clang 21.1.2 while gcc-15.2 has now problems with the new kernel 16.17.1  and new nvidia drivers... 

I Tried to mitigate a newly occuring bug with collapse(2) in a teams distribute parallel for construct in gcc for matrix multiply that occurs with the new 6.17 kernel and the gpu drivers nvidia-drivers-580.95.05. Unfortunately, with the new kernel 6.17.1 and the new gpu drivers nvidia-drivers-580.95.05, there seem to occur memory problems that do not appear to come from my code but incorrect compilation and libgomp. 

However, after the last fixes, https://releases.llvm.org/21.1.0/tools/clang/docs/ReleaseNotes.html#openmp-support the library compiles and runs with clang 21.1.2, kernel 6.17.1 and nvidia-drivers nvidia-drivers-580.95.05